+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/CommunityId.kt
@@ -1,3 +1,3 @@
 package it.vercruysse.lemmyapi.v0x19.datatypes
 
-typealias CommunityId = Int
+typealias CommunityId = Long
