+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/GetPosts.kt
@@ -8,8 +8,8 @@ import kotlinx.serialization.Serializable
 data class GetPosts(
     val type_: ListingType? /* "All" | "Local" | "Subscribed" | "ModeratorView" */ = null,
     val sort: SortType? /* "Active" | "Hot" | "New" | "Old" | "TopDay" | "TopWeek" | "TopMonth" | "TopYear" | "TopAll" | "MostComments" | "NewComments" | "TopHour" | "TopSixHour" | "TopTwelveHour" | "TopThreeMonths" | "TopSixMonths" | "TopNineMonths" | "Controversial" | "Scaled" */ = null,
-    val page: Int? = null,
-    val limit: Int? = null,
+    val page: Long? = null,
+    val limit: Long? = null,
     val community_id: CommunityId? = null,
     val community_name: String? = null,
     val saved_only: Boolean? = null,
