+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/LocalSiteId.kt
@@ -1,3 +1,3 @@
 package it.vercruysse.lemmyapi.v0x19.datatypes
 
-typealias LocalSiteId = Int
+typealias LocalSiteId = Long
