+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/LocalUserId.kt
@@ -1,3 +1,3 @@
 package it.vercruysse.lemmyapi.v0x19.datatypes
 
-typealias LocalUserId = Int
+typealias LocalUserId = Long
