+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/PostReportId.kt
@@ -1,3 +1,3 @@
 package it.vercruysse.lemmyapi.v0x19.datatypes
 
-typealias PostReportId = Int
+typealias PostReportId = Long
